+++ packages/@magic-sdk/provider/test/spec/modules/auth/loginWithEmailOTP.spec.ts
@@ -37,27 +37,6 @@ test('Generates JSON RPC request payload with `showUI: false` parameter', async 
   expect(requestPayload.params).toEqual([{ email: expectedEmail, showUI: false }]);
 });
 
-test('Generates JSON RPC pending for otp-input-sent', async () => {
-  const magic = createMagicSDK();
-  magic.auth.overlay.post = jest.fn();
-  const createIntermediaryEventFn = jest.fn();
-  magic.auth.createIntermediaryEvent = jest.fn().mockImplementation(() => createIntermediaryEventFn);
-
-  const handle = magic.auth.loginWithEmailOTP({ email: expectedEmail, showUI: false });
-
-  const troll_otp = '123456';
-  handle.emit('verify-email-otp', troll_otp);
-  handle.emit('cancel');
-
-  const verifyEvent = magic.auth.createIntermediaryEvent.mock.calls[0];
-  expect(verifyEvent[0]).toBe('verify-email-otp');
-  expect(verifyEvent[1]).toBe(3); // third call of test
-  expect(createIntermediaryEventFn.mock.calls[0][0]).toBe(troll_otp);
-
-  const intermediaryEventSecondMethod = magic.auth.createIntermediaryEvent.mock.calls[1][0];
-  expect(intermediaryEventSecondMethod).toBe('cancel');
-});
-
 test('If `testMode` is enabled, testing-specific RPC method is used', async () => {
   const magic = createMagicSDKTestMode();
   magic.auth.request = jest.fn();
